+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 ## Added
 
+- `options` parameter in `getSender` method to modify W5 and Highload V3 wallets
+
+- `customLiteSequencerEndpoints` parameter in `SDKParams` to specify custom lite sequencer endpoints
+
 - `simulateEVMMessage` method in `TacSdk` to simulate EVM message execution on TAC side
 
     ```typescript